test: completely isolate Playwright UI tests from backend unit test directory
