+++ .github/workflows/ui-tests.yml
@@ -27,4 +27,4 @@ jobs:
       run: playwright install chromium --with-deps
         
     - name: Run Playwright Tests
-      run: pytest tests/ui/ --browser chromium
+      run: pytest ui_tests/ --browser chromium
